+++ eras/conway/impl/src/Cardano/Ledger/Conway/Rules/Deleg.hs
@@ -20,12 +20,22 @@ module Cardano.Ledger.Conway.Rules.Deleg (
   ConwayDelegPredFailure (..),
   ConwayDelegEnv (..),
   processDelegation,
+  -- For testing only
+  incorrectDeposit,
+  incorrectRefund,
 ) where
 
-import Cardano.Ledger.BaseTypes (ProtVer (..), ShelleyBase, StrictMaybe (..), natVersion)
+import Cardano.Ledger.BaseTypes (
+  Mismatch (..),
+  ProtVer (..),
+  Relation (RelEQ),
+  ShelleyBase,
+  StrictMaybe (..),
+  natVersion,
+ )
 import Cardano.Ledger.Binary (DecCBOR (..), EncCBOR (..))
 import Cardano.Ledger.Binary.Coders (
-  Decode (From, Invalid, SumD, Summands),
+  Decode (From, FromGroup, Invalid, SumD, Summands),
   Encode (..),
   decode,
   encode,
@@ -35,7 +45,12 @@ import Cardano.Ledger.Binary.Coders (
 import Cardano.Ledger.Coin (Coin, compactCoinOrError)
 import Cardano.Ledger.Compactible (fromCompact)
 import Cardano.Ledger.Conway.Core
-import Cardano.Ledger.Conway.Era (ConwayDELEG, ConwayEra, hardforkConwayBootstrapPhase)
+import Cardano.Ledger.Conway.Era (
+  ConwayDELEG,
+  ConwayEra,
+  hardforkConwayBootstrapPhase,
+  hardforkConwayDELEGIncorrectDepositsAndRefunds,
+ )
 import Cardano.Ledger.Conway.State
 import Cardano.Ledger.Conway.TxCert (
   ConwayDelegCert (ConwayDelegCert, ConwayRegCert, ConwayRegDelegCert, ConwayUnRegCert),
@@ -96,6 +111,8 @@ data ConwayDelegPredFailure era
   | StakeKeyHasNonZeroRewardAccountBalanceDELEG Coin
   | DelegateeDRepNotRegisteredDELEG (Credential 'DRepRole)
   | DelegateeStakePoolNotRegisteredDELEG (KeyHash 'StakePool)
+  | DepositIncorrectDELEG (Mismatch 'RelEQ Coin)
+  | RefundIncorrectDELEG (Mismatch 'RelEQ Coin)
   deriving (Show, Eq, Generic)
 
 type instance EraRuleFailure "DELEG" ConwayEra = ConwayDelegPredFailure ConwayEra
@@ -123,6 +140,10 @@ instance Era era => EncCBOR (ConwayDelegPredFailure era) where
         Sum (DelegateeDRepNotRegisteredDELEG @era) 5 !> To delegatee
       DelegateeStakePoolNotRegisteredDELEG delegatee ->
         Sum (DelegateeStakePoolNotRegisteredDELEG @era) 6 !> To delegatee
+      DepositIncorrectDELEG mm ->
+        Sum (DepositIncorrectDELEG @era) 7 !> ToGroup mm
+      RefundIncorrectDELEG mm ->
+        Sum (RefundIncorrectDELEG @era) 8 !> ToGroup mm
 
 instance Era era => DecCBOR (ConwayDelegPredFailure era) where
   decCBOR = decode $ Summands "ConwayDelegPredFailure" $ \case
@@ -132,6 +153,8 @@ instance Era era => DecCBOR (ConwayDelegPredFailure era) where
     4 -> SumD StakeKeyHasNonZeroRewardAccountBalanceDELEG <! From
     5 -> SumD DelegateeDRepNotRegisteredDELEG <! From
     6 -> SumD DelegateeStakePoolNotRegisteredDELEG <! From
+    7 -> SumD DepositIncorrectDELEG <! FromGroup
+    8 -> SumD RefundIncorrectDELEG <! FromGroup
     n -> Invalid n
 
 instance
@@ -169,7 +192,7 @@ conwayDelegTransition = do
     ppKeyDepositCompact = compactCoinOrError ppKeyDeposit
     pv = pp ^. ppProtocolVersionL
     checkDepositAgainstPParams deposit =
-      deposit == ppKeyDeposit ?! IncorrectDepositDELEG deposit
+      deposit == ppKeyDeposit ?! incorrectDeposit pv deposit ppKeyDeposit
     checkStakeKeyNotRegistered stakeCred =
       not (isAccountRegistered stakeCred accounts) ?! StakeKeyRegisteredDELEG stakeCred
     checkStakeKeyIsRegistered stakeCred = do
@@ -207,14 +230,15 @@ conwayDelegTransition = do
             -- we don't want to report invalid refund when stake credential is not registered:
             accountState <- mAccountState
             -- we return offending refund only when it doesn't match the expected one:
-            guard (suppliedRefund /= fromCompact (accountState ^. depositAccountStateL))
-            Just suppliedRefund
+            let expectedRefund = fromCompact $ accountState ^. depositAccountStateL
+            guard (suppliedRefund /= expectedRefund)
+            Just $ incorrectRefund pv suppliedRefund expectedRefund
           checkStakeKeyHasZeroRewardBalance = do
             accountState <- mAccountState
             let balanceCompact = accountState ^. balanceAccountStateL
             guard (balanceCompact /= mempty)
             Just $ fromCompact balanceCompact
-      failOnJust checkInvalidRefund IncorrectDepositDELEG
+      failOnJust checkInvalidRefund id
       isJust mAccountState ?! StakeKeyNotRegisteredDELEG stakeCred
       failOnJust checkStakeKeyHasZeroRewardBalance StakeKeyHasNonZeroRewardAccountBalanceDELEG
       pure $
@@ -309,3 +333,25 @@ processDRepUnDelegation stakeCred (Just delegatee) cState =
               dRepState {drepDelegs = Set.delete stakeCred (drepDelegs dRepState)}
          in vState & vsDRepsL %~ Map.adjust removeDelegation dRepCred
       _ -> vState
+
+incorrectDeposit :: ProtVer -> Coin -> Coin -> ConwayDelegPredFailure era
+incorrectDeposit pv supplied expected =
+  if hardforkConwayDELEGIncorrectDepositsAndRefunds pv
+    then
+      DepositIncorrectDELEG
+        Mismatch
+          { mismatchSupplied = supplied
+          , mismatchExpected = expected
+          }
+    else IncorrectDepositDELEG supplied
+
+incorrectRefund :: ProtVer -> Coin -> Coin -> ConwayDelegPredFailure era
+incorrectRefund pv supplied expected =
+  if hardforkConwayDELEGIncorrectDepositsAndRefunds pv
+    then
+      RefundIncorrectDELEG
+        Mismatch
+          { mismatchSupplied = supplied
+          , mismatchExpected = expected
+          }
+    else IncorrectDepositDELEG supplied